pyproject.toml
[project]
name = "desk_controller"
version = "0.1.0"
description = "Desk Controller V1 - Python implementation of Desk Controller Project"
readme = "README.md"
authors = [
  {name = "Nathan Coates"},
]
requires-python = ">=3.13"
dependencies = [
    "gpiozero>=2.0.1",
    "pillow>=11.1.0",
    "pyserial>=3.5",
    "python-dotenv>=1.0.1",
    "python-hue-v2>=2.2.1",
    "rpi-lgpio>=0.6",
    "schedule>=1.2.2",
    "smbus>=1.1.post2",
    "spidev>=3.6",
    "spotipy>=2.25.1",
]

[dependency-groups]
dev = [
    "pygame>=2.6.1",
    "ruff>=0.11.2",
]
